new init features
added print_block function to make longer menus easier to read/write, I think this could make some of the plotting easier to code. Also updated the character generator function to implement new classes, and clean up the style a little. This is a little of my taste, so if I need to revert anything to be more in line with the group let me know.