print formatter
This is a function I came up with elsewhere in my studies. It's takes a line width and formats the input string so that it adds a hyphen if it splits a word at the end of a line, and does a couple checks for tidiness. I was thinking about the json plots, this might help from having to format the text too much on that end. It could still use a little help so that it formats the output like __init__.py does(start each line with '>'). It's also something that can probably wait until we're either working with a fixed window size or grabbing the details from the user to ensure consistent printing.